Implement limit checking for CloudFront service
Related to issue #539

List of implemented limits:
- Alternate domain names (CNAMEs) per distribution
- Cache behaviors per distribution
- Origin groups per distribution
- Key groups associated with a single distribution
- Origins per distribution
- Whitelisted cookies per cache behavior
- Whitelisted headers per cache behavior
- Whitelisted query strings per cache behavior
- Key groups associated with a single cache behavior
- Public keys in a single key group
- Distributions associated with a single key group
- Cookies per cache policy
- Headers per cache policy
- Query strings per cache policy
- Distributions associated with the same cache policy
- Cookies per origin request policy
- Headers per origin request policy
- Query strings per origin request policy
- Distributions associated with the same origin request policy

Unit tests implemented.

The CloudFront service is considered as global, so the same usage will be reported for CloudFront, no matter which region we are using.

Note: in the "Service Quotas" service, the CloudFront quotas are listed in the region "us-east-1".

The AWS "Service Quotas" service does not return the applied service quotas for CloudFront, only the default quotas.