Replace "Max spot instance requests per region" limit with vCPU-based limits
BREAKING CHANGE

As mentioned in issue #502, the limit "Max spot instance requests per region" was removed from AWS and replaced by the following limits, based on the number of vCPU:
- All F Spot Instance Requests
- All G Spot Instance Requests
- All Inf Spot Instance Requests
- All P Spot Instance Requests
- All Standard (A, C, D, H, I, M, R, T, Z) Spot Instance Requests
- All X Spot Instance Requests

As a result, the usage reported by awslimitchecker is not relevant anymore.

In order to fix this, this commit replaces the call to the EC2 API describe_spot_instance_requests() with a call to the CloudWatch API, since the usage of Spot requests vCPUs is now available as a metric in CloudWatch.

In addition, the applied limits are now retrieved from the Service Quotas service.